Revision exhaustiva de las clases Pt.2

+++ src/main/java/com/example/proydbp/pedido_local/dto/PedidoLocalResponseDto.java
@@ -28,7 +28,7 @@ public class PedidoLocalResponseDto {
 
     private OrderResponseDto orden;
 
-    private BigDecimal precio;
+    private Double precio;
 
     private TipoPago tipoPago;
 }
